[#99] Added LevelBuilding that is extended by buildings with levels
* Created LevelBuilding that declares state and all functions previously defined by BoydOrr and Library
* Extends BoydOrr and Library
